restructure: move src/ to core/, add gfx module with SDL3+GL bootstrap
- Rename src/ -> core/ to make room for optional gfx/gui modules
- Add gfx/ with pl_gfx_init/quit: SDL3 window + OpenGL context creation,
  Wayland/X11 detection, driver/vendor/renderer logging via glad
- Add include/pl_gfx.h, pl_gfx_gl.h (internal), pl_gui.h (stub for future gui module)
- Add include/pl_sys.h
- Update Stormfile: gate gfx/gui builds behind --set=gfx/--set=gui,
  single static target (plenum-debug) with conditional sources
- Add test/test_gfx.c: verifies core init guard + real GL context
  creation, driver detection end to end
- Add test/test_thread.c
- Minor fixes to pl_cpu.h, pl_defs.h, pl_thread.h, existing tests

gfx currently supports GL backend only;

+++ core/pl_cpu.c
@@ -1,15 +1,15 @@
 #include "pl_cpu.h"
 #include <unistd.h>
 
-u32 pl_cpu_nproc(void)
+u64 pl_cpu_nproc(void)
 {
     i64 n = sysconf(_SC_NPROCESSORS_ONLN);
 
-    return n < 1 ? 1 : (u32) n;
+    return n < 1 ? 1 : (u64) n;
 }
 
-u32 pl_cpu_cacheline(void)
+u64 pl_cpu_cacheline(void)
 {
     i64 size = sysconf(_SC_LEVEL1_DCACHE_LINESIZE);
-    return (size <= 0) ? 64 : (u32) size;
+    return (size <= 0) ? 64 : (u64) size;
 }
